Restructured directoy for 'mvn test' utilization
