+++ pom.xml
@@ -6,7 +6,7 @@
   <name>UMLEditor</name>
   <description>UMLEditor for CSCI420</description>
   <build>
-    <sourceDirectory>src</sourceDirectory>
+    <sourceDirectory>src/main/java</sourceDirectory>
     <plugins>
       <plugin>
         <artifactId>maven-compiler-plugin</artifactId>
@@ -15,6 +15,11 @@
           <source>1.8</source>
           <target>1.8</target>
         </configuration>
+      </plugin>
+      <plugin>
+      	<groupId>org.apache.maven.plugins</groupId>
+      	<artifactId>maven-sunfire-plugin</artifactId>
+      	<version>2.22.0</version>
       </plugin>
     </plugins>
   </build>
