+++ src/main/java/core/UMLFileIO.java
@@ -1,4 +1,4 @@
-package main;
+package core;
 
 import java.io.File;
 import java.io.FileNotFoundException;
